+++ secretflow/component/io/core/bins/__init__.py
@@ -4,12 +4,14 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#   http://www.apache.org/licenses/LICENSE-2.0
+#      https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.∏
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = "1.3.0.dev20231201"
+# This folder contains how to serialize model to custom pb and back to a new model.
+
+# complete_model -> partial_info_pb + hidden_old_model -> partial_info_pb (modified) + hidden_old_model -> complete_model (new)