+++ secretflow/component/ml/linear/ss_sgd.py
@@ -276,7 +276,7 @@ ss_sgd_predict_comp.bool_attr(
     ),
     is_list=False,
     is_optional=True,
-    default_value=False,
+    default_value=True,
 )
 ss_sgd_predict_comp.bool_attr(
     name="save_label",